Added prototype of multi_conversion()
Added the first version of multi_conversion() but it needs to be tested and also it just returns a list of tuples, with each tuple containing a circuit and an i/o map dictionary. These have to be merged together into a single new circuit which it doesn't do yet.